requirements.txt
pandas
pandas-datareader
numpy
alpaca-trade-api
python-dotenv
requests
statsmodels
flask